# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Library Management API",
        "description": "API for managing library operations including books, users, and borrowing.",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/books": {
            "get": {
                "tags": [
                    "Books"
                ],
                "summary": "Get list of books",
                "operationId": "8e67fc126a51598255c56d362d50466e",
                "responses": {
                    "200": {
                        "description": "Books retrieved successfully"
                    }
                }
            },
            "post": {
                "tags": [
                    "Books"
                ],
                "summary": "Create a new book",
                "operationId": "6c5e00fe01c5e356fa95e728c3cfbc98",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "author",
                                    "published_at",
                                    "isbn"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "The Lord of the Rings"
                                    },
                                    "author": {
                                        "type": "string",
                                        "example": "J.R.R. Tolkien"
                                    },
                                    "published_at": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "1954-07-29"
                                    },
                                    "isbn": {
                                        "type": "string",
                                        "example": "978-0-618-64015-7"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "A fantasy novel."
                                    },
                                    "cover_image": {
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "stock": {
                                        "type": "integer",
                                        "example": 10
                                    },
                                    "categories": {
                                        "type": "array",
                                        "items": {
                                            "type": "integer",
                                            "example": 1
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Book created successfully"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Unauthorized"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/books/{book}": {
            "get": {
                "tags": [
                    "Books"
                ],
                "summary": "Get book details",
                "operationId": "4fd937759f65ce4d38eeee7937a93a11",
                "parameters": [
                    {
                        "name": "book",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Book retrieved successfully"
                    },
                    "404": {
                        "description": "Book not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Books"
                ],
                "summary": "Update a book",
                "operationId": "7e2dea68c5c8e98bc212b57acea73f06",
                "parameters": [
                    {
                        "name": "book",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "The Lord of the Rings"
                                    },
                                    "author": {
                                        "type": "string",
                                        "example": "J.R.R. Tolkien"
                                    },
                                    "published_at": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "1954-07-29"
                                    },
                                    "isbn": {
                                        "type": "string",
                                        "example": "978-0-618-64015-7"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "A fantasy novel."
                                    },
                                    "cover_image": {
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "stock": {
                                        "type": "integer",
                                        "example": 10
                                    },
                                    "categories": {
                                        "type": "array",
                                        "items": {
                                            "type": "integer",
                                            "example": 1
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Book updated successfully"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Book not found"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Books"
                ],
                "summary": "Delete a book",
                "operationId": "1870d5221f34fa03e3ece095b7bf2b63",
                "parameters": [
                    {
                        "name": "book",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Book deleted successfully"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Book not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/books/search": {
            "get": {
                "tags": [
                    "Books"
                ],
                "summary": "Search for books",
                "operationId": "1b5f924a84be4bd1ca201ed19c175911",
                "parameters": [
                    {
                        "name": "q",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Books retrieved successfully"
                    }
                }
            }
        },
...
    },
...
}